Move map context popup to TSX
